Add files via upload
Hi Dr. Im, 
I understand this is not precisely what you requested and I wasn't able to finish it. I was testing the code first before 
implementing it in my project. I couldn't get the deserialization, it wasn't reading. I will finish it anyways as a personal project to learn. 

Thank you for your help!

Jose